+++ BinanceExchange.API/Models/Request/CreateOrderRequest.cs
@@ -31,7 +31,7 @@ namespace BinanceExchange.API.Models.Request
         public decimal Quantity { get; set; }
 
         [DataMember(Order = 6)]
-        public decimal Price { get; set; }
+        public decimal? Price { get; set; }
 
         [DataMember(Order = 7)]
         public string NewClientOrderId { get; set; }
@@ -42,5 +42,10 @@ namespace BinanceExchange.API.Models.Request
         [DataMember(Order = 9)]
         [JsonProperty("icebergQty")]
         public decimal? IcebergQuantity { get; set; }
+
+        [DataMember(Order = 10)]
+        [JsonProperty("newOrderRespType")]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public NewOrderResponseType NewOrderResponseType { get; set; }
     }
 }