+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # BinanceDotNet Changelog
 
+**Release Date: 1/30/2018**
+## 4.2.0
+- Adjusted `CreateOrder` request to allow you to specify what type of response you want
+- Removed `TimeInForce` as required param from New order
+- Changed `Quantity` to be `decimal`
+
 **Release Date: 1/23/2018**
 ## 4.0.1
 - Fixed bug around QueryOrder Http method